fix(example): next.js production runtime error
